LPS-72928 Move to configuration package to match other cases
